backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
langgraph
groq
python-dotenv
